pyproject.toml
[project]
name = "beholder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.8.5",
    "googlemaps>=4.10.0",
    "ipykernel>=6.30.0",
    "ipython>=9.4.0",
    "matplotlib>=3.10.5",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
